@@ -3,10 +3,20 @@
 namespace App\Entity;
 
 use App\Repository\SkillRepository;
+use DateTime;
+use DateTimeInterface;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\HttpFoundation\File\File;
+use Vich\UploaderBundle\Mapping\Annotation as Vich;
+use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Form\Extension\Core\Type\FileType;
+use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+
+
 
 /**
  * @ORM\Entity(repositoryClass=SkillRepository::class)
+ * @Vich\Uploadable
  */
 class Skill
 {
@@ -15,17 +25,31 @@ class Skill
      * @ORM\GeneratedValue
      * @ORM\Column(type="integer")
      */
-    private $id;
+    private ?int $id;
 
     /**
      * @ORM\Column(type="string", length=255)
      */
-    private $name;
+    private ?string $name;
 
     /**
-     * @ORM\Column(type="text")
+     * @ORM\Column(type="text", nullable=true)
      */
-    private $logo;
+    private ?string $logo='';
+
+    /**
+     * @Vich\UploadableField(mapping="skill_logo", fileNameProperty="logo")
+     * @var File|null
+     * @Assert\File(
+     *     maxSize="1000000",
+     *     mimeTypes = {"image/png", "image/jpeg", "image/jpg", "image/gif",})
+     */
+    private ?File $skillLogo = null;
+
+    /**
+     * @ORM\Column(type="datetime", nullable=true)
+     */
+    private ?DateTimeInterface $updatedAt;
 
     public function getId(): ?int
     {
@@ -54,5 +78,30 @@ class Skill
         $this->logo = $logo;
 
         return $this;
+    }
+
+    public function setSkillLogo(?File $image = null): Skill
+    {
+        $this->skillLogo = $image;
+        if ($image) {
+            $this->updatedAt = new DateTime('now');
+        }
+
+        return $this;
+    }
+
+    public function getSkillLogo(): ?File
+    {
+        return $this->skillLogo;
+    }
+
+    public function getUpdatedAt(): ?DateTimeInterface
+    {
+        return $this->updatedAt;
+    }
+
+    public function setUpdatedAt(?DateTimeInterface $updatedAt): self
+    {
+        $this->updatedAt = $updatedAt;
     }
 }
